llm-agent/requirements.txt

langchain
langchain-google-vertexai
langchain-qdrant
qdrant-client
google-cloud-aiplatform
google.cloud 